pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "story1-llm-project"
version = "0.1.0"
description = "Production-ready retrieval, reranking, and grounded QA API over BEIR subsets"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "OpenAI" }]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "sentence-transformers>=3.0.0,<4.0.0",
  "transformers>=4.44.0,<5.0.0",
  "torch>=2.4.0,<3.0.0",
  "faiss-cpu>=1.8.0.post1,<2.0.0",
  "rank-bm25>=0.2.2,<1.0.0",
  "ir-datasets>=0.5.8,<1.0.0",
  "orjson>=3.10.7,<4.0.0",
  "httpx>=0.27.0,<1.0.0",
  "openai>=1.43.0,<2.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "prometheus-client>=0.20.0,<1.0.0",
  "structlog>=24.4.0,<25.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "numpy>=1.26.4,<3.0.0",
  "tqdm>=4.66.5,<5.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "redis>=5.0.8,<6.0.0",
  "rq>=1.16.2,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.6.2,<1.0.0",
  "mypy>=1.11.1,<2.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
